Use the the standard BUILD_TESTING CMake variable
Use this as a cache variable to allow it to be overridden by users and other projects.
There are better practices for enabling / disabling tests in CMake.
It's normal to include a check for whether or not this project is top-level and disable building tests by default.
It's also recommended to provide a project-specific variable for controlling building tests or not.

I've kept this change minimal for now, however.
This makes integration with the Conan package manager easier.