Disable running tests on MSVC even with Bash and sed
The GitHub hosted runners for MSVC have Bash and apparently sed, too.
This causes tests to be built when using MSVC on Windows incorrectly.

This PR moves the check for MSVC which disables tests before the check for Bash and sed.
This ensures that the CMake cache variable is set properly for MSVC.