Fix test execution by complete refactoring.
Previously, the test scripts have been executed from a rewritten version, which
used `sed` to adjust binary paths. Not only have the tests passed despite the
scripts not being created correctly, this mechanism was also cumbersome and
intransparent.

This refactoring now uses the original test scripts from the source tree.

When building/testing with `cmake`/`ctest`, the (absolute) paths to executables
are being taken from the environment, which is set in CMake via the new
`declare_test_executable()` function.

When executing tests from `make check` (or for manual execution), the current
dir is used. Which is fine, if executed from the cmake build dir.

The path to the files containing the expected data is passed as parameter now.

Furthermore, some optimisations have been done in the bash scripts. Output files
of the test execution are not deleted anymore, which should be OK, if executed
in the build directory. EG_DATA is now also being passed as parameter.

Refactor and add contrib test also, although not clear how this is connected
and how this adds value to contrib...